requirements-dev.txt
coverage
django-test-plus==1.4.0
factory-boy
flake8
mypy==0.790
pip-tools==5.3.1
pre-commit
pygraphviz==1.6
pytest-django==3.10.0
